feat: add saving auth data
