+++ blackjack-classic/src/hocs/WithEventListener/index.ts
@@ -1,0 +1,1 @@
+export { WithEventListener } from './WithEventListener';
