+++ blackjack-classic/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './authorization';
+export * from './dealer';
