+++ blackjack-classic/src/store/index.ts
@@ -1,1 +1,3 @@
 export * from './store';
+export * from './game';
+export * from './user';
