+++ blackjack-classic/src/store/middleware/index.ts
@@ -1,0 +1,1 @@
+export { saveUserDataMiddleware } from './middleware';
